+++ src/main/java/org/joelson/turf/scorecalc/imprt/model/ZoneImportRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ZoneImportRepository extends JpaRepository<ZoneImport, Long> {
+public interface ZoneImportRepository extends JpaRepository<ZoneImport, Integer> {
 
-    <T> Optional<T> findByZoneId(Long zoneId, Class<T> type);
+    <T> Optional<T> findByZoneId(Integer zoneId, Class<T> type);
 }
